# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Mishwar - Complete ride hailing and group trips platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /driver-bookings:
    get:
      operationId: listDriverBookings
      tags: [driverBookings]
      summary: List my driver bookings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverBookingList"
    post:
      operationId: createDriverBooking
      tags: [driverBookings]
      summary: Book a private driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDriverBookingBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverBooking"
  /driver-bookings/{bookingId}:
    get:
      operationId: getDriverBooking
      tags: [driverBookings]
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverBooking"
    patch:
      operationId: updateDriverBooking
      tags: [driverBookings]
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDriverBookingBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverBooking"

  # Driver Portal
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [passenger, driver, admin]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, banned]
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserList"
  /admin/users/{userId}/status:
    patch:
      operationId: updateUserStatus
      tags: [admin]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserStatusBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/drivers:
    get:
      operationId: listAdminDrivers
      tags: [admin]
      summary: List all drivers
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, suspended]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDriverList"
  /admin/drivers/{driverId}/status:
    patch:
      operationId: updateDriverStatus
      tags: [admin]
      parameters:
        - name: driverId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDriverStatusBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /admin/trips:
    get:
      operationId: listAdminTrips
      tags: [admin]
      summary: List all trips/rides
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideList"
  /admin/group-trips:
    get:
      operationId: listAdminGroupTrips
      tags: [admin]
      summary: List all group trips
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupTripList"

  # Stats
...
components:
  schemas:
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [passenger, driver, admin]
          default: passenger
      required: [name, phone, password]
    LoginBody:
      type: object
      properties:
        phone:
          type: string
        password:
          type: string
      required: [phone, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [passenger, driver, admin]
        avatarUrl:
          type: string
          nullable: true
        rating:
          type: number
          nullable: true
        status:
          type: string
          enum: [active, suspended, banned]
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, phone, role, createdAt]
    Driver:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        phone:
          type: string
        vehicleMake:
          type: string
        vehicleModel:
          type: string
        vehicleYear:
          type: integer
        vehiclePlate:
          type: string
        vehicleType:
          type: string
          enum: [economy, comfort, xl, bus, van, minibus]
        rating:
          type: number
          nullable: true
        totalTrips:
          type: integer
        isAvailable:
          type: boolean
        driverStatus:
          type: string
          enum: [pending, approved, rejected, suspended]
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, name, phone, vehicleMake, vehicleModel, vehicleYear, vehiclePlate, vehicleType, totalTrips, isAvailable, createdAt]
...
    Ride:
      type: object
      properties:
        id:
          type: integer
        passengerId:
          type: integer
        driverId:
          type: integer
          nullable: true
        fromAddress:
          type: string
        toAddress:
          type: string
        fromLat:
          type: number
          nullable: true
        fromLng:
          type: number
          nullable: true
        toLat:
          type: number
          nullable: true
        toLng:
          type: number
          nullable: true
        status:
          type: string
          enum: [pending, accepted, in_progress, completed, cancelled]
        estimatedPrice:
          type: number
          nullable: true
        finalPrice:
          type: number
          nullable: true
        vehicleType:
          type: string
          enum: [economy, comfort, xl]
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        driver:
          $ref: "#/components/schemas/Driver"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, passengerId, fromAddress, toAddress, status, vehicleType, createdAt]
    RideList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Ride"
        total:
          type: integer
      required: [items, total]
...
    GroupTrip:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        fromCity:
          type: string
        toCity:
          type: string
        fromAddress:
          type: string
        toAddress:
          type: string
        departureTime:
          type: string
        scheduleType:
          type: string
          enum: [weekly, monthly, one_time]
        scheduleDays:
          type: string
          nullable: true
        scheduleDay:
          type: integer
          nullable: true
        capacity:
          type: integer
        registeredCount:
          type: integer
        pricePerSeat:
          type: number
        vehicleType:
          type: string
          enum: [bus, van, minibus]
        driverId:
          type: integer
          nullable: true
        driver:
          $ref: "#/components/schemas/Driver"
          nullable: true
        isActive:
          type: boolean
        nextDepartureDate:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, fromCity, toCity, fromAddress, toAddress, departureTime, scheduleType, capacity, registeredCount, pricePerSeat, vehicleType, isActive, createdAt]
    GroupTripList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/GroupTrip"
        total:
          type: integer
      required: [items, total]
...
    DriverBooking:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        driverId:
          type: integer
          nullable: true
        purpose:
          type: string
        startDatetime:
          type: string
          format: date-time
        endDatetime:
          type: string
          format: date-time
        location:
          type: string
        totalHours:
          type: number
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, confirmed, active, completed, cancelled]
        driver:
          $ref: "#/components/schemas/Driver"
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, userId, purpose, startDatetime, endDatetime, location, totalHours, totalPrice, status, createdAt]
    DriverBookingList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/DriverBooking"
        total:
          type: integer
      required: [items, total]
    CreateDriverBookingBody:
      type: object
      properties:
        driverId:
          type: integer
        purpose:
          type: string
        startDatetime:
          type: string
          format: date-time
        endDatetime:
          type: string
          format: date-time
        location:
          type: string
        notes:
          type: string
      required: [purpose, startDatetime, endDatetime, location]
    UpdateDriverBookingBody:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, active, completed, cancelled]
        driverId:
          type: integer

    # Driver Portal
...
    AdminUserList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
      required: [items, total]
    AdminDriverList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Driver"
        total:
          type: integer
      required: [items, total]
    UpdateUserStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [active, suspended, banned]
      required: [status]
    UpdateDriverStatusBody:
      type: object
      properties:
        driverStatus:
          type: string
          enum: [approved, rejected, suspended]
        reason:
          type: string
      required: [driverStatus]

    # Stats
...